moved functions into separate file
